Create dummy system web server code
In this code, a dummy system was created to test making a website that has functionalities that interact with hardware components. Here, you can see there is a two-light system that alternates lights, and the resulting website continuously logs the status of those lights, with updates. This served as a good base to develop website code with a more robust hardware system associated with it.