Create Functional Code without Website
This code was the integration of the LEDs, Ultrasonic Sensors, and the Motor - Button systems. A primary change that occurred here was that now the motor turns back the amount that it turns forward to begin with - ensuring essentially that the door closed after it opened. 

The code from the ultrasonic branch was merged with the code from the stepper motor branch, such that the the ultrasonic sensors turned off once the stepper motor was done moving. 

LEDs were also integrated into this section of the code, such that, with dummy variables the code for distances, specific LEDs turned on for specified distance ranges that the ultrasonic sensors detected.