Create FINAL integrated code with website signals
This code change now includes website compatibility. It does this with a signal called INDICATOR that turns on when the website should poll the system for data. The code for the actual website is in a different file because we had to use another Arduino to run the website to stop it from completely timing out when we had a delay in the code after the red light was turned on. This ended up shutting down the website, so instead we used a second Arduino to host the website, as suggested by Professor Asare.