Create Initial Webstie On Base Code
In this code we tried to just make a simple website that had one line of text with the other code that controls the rest of the logic for the circuit. However, we discovered that this website crashed entirely when the code locks the door when the light turns red. For this reason, we have to try something else.