Create Final Website Code
This code is the final code that was used for the website. We ended up having to run this code on a second Arduino and connect it to the main Arduino that runs the logic code for the rest of the box. Note that the date and time are hardcoded here this is just because we needed extra modules to have that working, which we did not deem necessary for proof of concept of the prototype.