Create FINAL code for website on second arduino
same as the code in the website development branch. This is because it is exactly the same code but it should all be on the main branch, so it is. This runs on the second arduino the other file with FINAL in it is for the first arduino that sits inside the box.